+++ handler/location/location.go
@@ -3,8 +3,8 @@ package location
 import "indoor_positioning/model"
 
 type Fingerprint struct {
-	Bssid string `json:"bssid"`
-	Rss   int    `json:"rss"`
+	Bssid string  `json:"bssid"`
+	Rss   float64 `json:"rss"`
 }
 
 // location相关api所需请求响应结构
@@ -17,9 +17,9 @@ type GetResponse struct {
 }
 
 type CreateKnnRequest struct {
-	Offline_rss      [][]int     `json:"offline_rss"`
+	Offline_rss      [][]float64 `json:"offline_rss"`
 	Offline_location [][]float64 `json:"offline_location"`
-	Online_rss       [][]int     `json:"online_rss"`
+	Online_rss       [][]float64 `json:"online_rss"`
 }
 type CreateKnnResponse struct {
 	Coordinate [][]float64 `json:"coordinate"`
